+++ Tests/SentryTests/SentryOptionsTest.m
@@ -1291,6 +1291,27 @@
     [self testBooleanField:@"enableSpanLocalMetricAggregation" defaultValue:YES];
 }
 
+- (void)testBeforeEmitMetric
+{
+    SentryBeforeEmitMetricCallback callback
+        = ^(NSString *_Nonnull key, NSDictionary<NSString *, NSString *> *_Nonnull tags) {
+              // Use tags and key to silence unused compiler error
+              XCTAssertNotNil(key);
+              XCTAssertNotNil(tags);
+              return YES;
+          };
+    SentryOptions *options = [self getValidOptions:@{ @"beforeEmitMetric" : callback }];
+
+    XCTAssertEqual(callback, options.beforeEmitMetric);
+}
+
+- (void)testDefaultBeforeEmitMetric
+{
+    SentryOptions *options = [self getValidOptions:@{}];
+
+    XCTAssertNil(options.beforeEmitMetric);
+}
+
 #pragma mark - Private
 
 - (void)assertArrayEquals:(NSArray<NSString *> *)expected actual:(NSArray<NSString *> *)actual